move WorkerLost to a library file so it can be imported properly
